1. Created a VerificationService object in the AuthenticationService class 2. Replaced authenticationService.verifyResetPasswordCode with verificationService.verifyResetPasswordCode 3. Removed verifyMfaVerificationCodeAndAuthenticateUser and verifyResetPasswordCode methods from AuthenticationService interface. 4. Removed declaration of mfaService and its dependency injections from AuthenticationService class 5. Replaced initializeAuthenticationAndContext for VerificationService.initializeAuthenticationAndContext() from static to instance method 6. Removed method verifyMfaVerificationCodeAndAuthenticateUser from AuthenticationService class 7. Converted getMemberDetails to verificationService.getMemberDetails 8. Removed verifyPasswordCode method from AuthenticationService class 9. Removed initializeAuthenticationAndContext method from AuthenticationService class 10. Removed clearMfaVerificationCodeSavedTemporarily from AuthenticationService class 11. Removed validateMfaVerificationOrOtpCode method from AuthenticationService class 12. Converted setContext static method to verificationService.setContext() instance method 13. Removed clearResetPasswordOtpSavedTemporarily method from AuthenticationService class 14. Converted findProfileToken to verification.findProfileToken, validateProfileTokenAndVerificationCode to verificatinoService.validateProfileTokenAndVerificationCode and verifyResetPasswordTokenHasNotExpired to verificationService.verifyResetPasswordTokenHasNotExpired 15. Removed findProfileToken method from AuthenticationService class. 16. Removed validateProfileTokenAndVerificationCode method from AuthenticationService class. 17. Removed verifyResetPasswordTokenHasNotExpired method from AuthenticationService class. 18. Removed getMemberDetails method from AuthenticationService class. 19. Created instance of VerificationService class and injected it. 20. Converted static method validateVerificationCode method to instance method verification.validateVerificationCode method 21. Converted authenticationService.verifyMfaVerificationCodeAuthenticationUser method to verificationService.verifyMfaVerificationCodeAndAuthenticateUser 22. Moved verifyMfaVerificationCodeAndAuthenticateUser, initializeAuthenticationAndContext and setContext methods to VerificationService interface. 23. Moved verifyMfaVerificationCodeAndAuthenticateUser method in VerificationServiceImpl class. 24. Moved validateMfaVerificationOrOtpCode method in VerificationServiceImpl class.
25. Moved initializeAuthenticationAndContext method in VerificationServiceImpl class.
26. Moved setContext method in VerificationServiceImpl class.
27. Moved setUserTimezoneAfterAuthentication method in VerificationServiceImpl class
28. Moved clearMfaVerificationCodeSavedTemporarily method in VerificationServiceImpl class
29. Moved saveAuthenticationTokensToRepositoryOrCache method to VerificationServiceImpl class
30. Moved verifyResetPasswordCode method to VerificationServiceImpl class
31. Moved getMemberDetails method to VerificationServiceImpl class
32. Moved validateProfileTokenAndResetPasswordCode method to VerificationServiceImpl class
33. Moved findProfileToken method to VerificationServiceImpl class
34. Moved verifyResetPasswordTokenHasNotExpired method to VerificationServiceImpl class
35. Moved validateProfileTokenAndVerificationCode method to VerificationServiceImpl class
36. Moved clearResetPasswordOtpSavedTemporarily method to VerificationServiceImpl class
