+++ src/main/java/com/fleencorp/feen/controller/security/AuthenticationController.java
@@ -10,6 +10,7 @@ import com.fleencorp.feen.model.response.auth.SignUpResponse;
 import com.fleencorp.feen.model.response.security.ForgotPasswordResponse;
 import com.fleencorp.feen.model.response.security.InitiatePasswordChangeResponse;
 import com.fleencorp.feen.service.auth.AuthenticationService;
+import com.fleencorp.feen.service.verification.VerificationService;
 import jakarta.validation.Valid;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.web.bind.annotation.*;
@@ -19,9 +20,11 @@ import org.springframework.web.bind.annotation.*;
 public class AuthenticationController {
 
   private final AuthenticationService authenticationService;
+  private final VerificationService verificationService;
 
-  public AuthenticationController(final AuthenticationService authenticationService) {
+  public AuthenticationController(final AuthenticationService authenticationService, final VerificationService verificationService) {
     this.authenticationService = authenticationService;
+    this.verificationService = verificationService;
   }
 
   @GetMapping(value = "/data-for-sign-up")
@@ -51,6 +54,6 @@ public class AuthenticationController {
   @PostMapping(value = "/verify-reset-password-code")
   public InitiatePasswordChangeResponse validateResetPasswordCode(
       @Valid @RequestBody final ResetPasswordDto resetPasswordDto) {
-    return authenticationService.verifyResetPasswordCode(resetPasswordDto);
+    return verificationService.verifyResetPasswordCode(resetPasswordDto);
   }
 }
