+++ src/main/java/com/fleencorp/feen/service/auth/AuthenticationService.java
@@ -1,13 +1,11 @@
 package com.fleencorp.feen.service.auth;
 
 import com.fleencorp.feen.model.dto.auth.*;
-import com.fleencorp.feen.model.dto.security.mfa.ConfirmMfaVerificationCodeDto;
 import com.fleencorp.feen.model.response.auth.DataForSignUpResponse;
 import com.fleencorp.feen.model.response.auth.SignInResponse;
 import com.fleencorp.feen.model.response.auth.SignUpResponse;
 import com.fleencorp.feen.model.response.security.ChangePasswordResponse;
 import com.fleencorp.feen.model.response.security.ForgotPasswordResponse;
-import com.fleencorp.feen.model.response.security.InitiatePasswordChangeResponse;
 import com.fleencorp.feen.model.response.security.SignOutResponse;
 import com.fleencorp.feen.model.security.FleenUser;
 
@@ -21,13 +19,9 @@ public interface AuthenticationService {
 
   SignOutResponse signOut(FleenUser user);
 
-  SignInResponse verifyMfaVerificationCodeAndAuthenticateUser(ConfirmMfaVerificationCodeDto confirmMfaVerificationCodeDto, FleenUser user);
-
   SignInResponse signIn(SignInDto signInDto);
 
   ForgotPasswordResponse forgotPassword(ForgotPasswordDto forgotPasswordDto);
-
-  InitiatePasswordChangeResponse verifyResetPasswordCode(ResetPasswordDto resetPasswordDto);
 
   ChangePasswordResponse changePassword(ChangePasswordDto changePasswordDto, FleenUser user);
 }
