+++ src/main/java/com/fleencorp/feen/service/impl/auth/AuthenticationServiceImpl.java
@@ -99,7 +99,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
   private final CacheService cacheService;
   private final CountryService countryService;
   private final MemberService memberService;
-  private final MfaService mfaService;
   private final RoleService roleService;
   private final TokenService tokenService;
   private final VerificationService verificationService;
@@ -125,7 +124,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
    * @param memberService the service for managing members and profile information
    * @param cacheService the service handling cache operations.
    * @param countryService the service providing country-related data and operations.
-   * @param mfaService the service managing multi-factor authentication (MFA).
    * @param roleService the service managing user roles.
    * @param tokenService the service handling token generation and validation.
    * @param verificationService the service for handling verifications like sign up and forgot password
@@ -157,7 +155,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     this.cacheService = cacheService;
     this.countryService = countryService;
     this.memberService = memberService;
-    this.mfaService = mfaService;
     this.roleService = roleService;
     this.tokenService = tokenService;
     this.verificationService = verificationService;
@@ -266,7 +263,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
    */
   protected FleenUser authenticateAndInitializeContext(final Member member) {
     // Initialize authentication and set context for the new member
-    final FleenUser user = initializeAuthenticationAndContext(member);
+    final FleenUser user = verificationService.initializeAuthenticationAndContext(member);
     // Set user timezone after authentication
     setUserTimezoneAfterAuthentication(user);
     // Return the user
@@ -380,7 +377,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     updateMemberProfile(member, verificationType);
 
     // Initialize authentication and context for the new user
-    final FleenUser newUser = initializeAuthenticationAndContext(member);
+    final FleenUser newUser = verificationService.initializeAuthenticationAndContext(member);
     // Clear temporary sign-up verification code
     clearSignUpVerificationCodeSavedTemporarily(username);
     // Set user timezone after authentication
@@ -525,39 +522,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
   }
 
   /**
-   * Verifies the Multi-Factor Authentication (MFA) verification code provided by the user,
-   * authenticates the user, and generates access and refresh tokens.
-   *
-   * @param confirmMfaCodeDto the DTO containing the verification code and MFA type to confirm
-   * @param user              the authenticated user performing the action
-   * @return SignInResponse containing the generated access and refresh tokens
-   * @throws InvalidAuthenticationException if the user cannot be found or authenticated
-   */
-  @Override
-  public SignInResponse verifyMfaVerificationCodeAndAuthenticateUser(final ConfirmMfaVerificationCodeDto confirmMfaCodeDto, final FleenUser user) {
-    final String username = user.getUsername();
-    final Member member = memberRepository.findByEmailAddress(username)
-      .orElseThrow(InvalidAuthenticationException.of(username));
-
-    // Validate the provided MFA verification code based on its type
-    validateMfaVerificationOrOtpCode(confirmMfaCodeDto.getVerificationCode(), confirmMfaCodeDto.getActualMfaType(), username, member.getMemberId());
-    // Initialize authentication and context for the authenticated user
-    final FleenUser authenticatedUser = initializeAuthenticationAndContext(member);
-    // Set user timezone after authentication
-    setUserTimezoneAfterAuthentication(authenticatedUser);
-    // Create access and refresh tokens for the authenticated user
-    final String accessToken = tokenService.createAccessToken(authenticatedUser, AuthenticationStatus.COMPLETED);
-    final String refreshToken = tokenService.createRefreshToken(authenticatedUser);
-
-    // Clear temporarily saved MFA verification code
-    clearMfaVerificationCodeSavedTemporarily(username);
-    // Save authentication tokens to repository or cache
-    saveAuthenticationTokensToRepositoryOrCache(username, accessToken, refreshToken);
-    // Return SignInResponse with access and refresh tokens
-    return localizedResponse.of(SignInResponse.of(accessToken, refreshToken));
-  }
-
-  /**
    * Signs in a user with the provided credentials.
    *
    * <p>This method authenticates the user using the provided email address and password.
@@ -693,7 +657,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Retrieve user's email address from DTO
     final String emailAddress = forgotPasswordDto.getEmailAddress();
     // Retrieve member details from repository or throw exception if not found
-    final Member member = getMemberDetails(emailAddress);
+    final Member member = verificationService.getMemberDetails(emailAddress);
     // Determine the verification type from DTO
     final VerificationType verificationType = forgotPasswordDto.getVerificationType();
     // Generate a random six-digit OTP as the reset password token.
@@ -712,37 +676,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
   }
 
   /**
-   * Validates the reset password code provided by the user and initiates the password change process.
-   *
-   * @param resetPasswordDto the DTO containing the reset password information
-   * @return an InitiatePasswordChangeResponse containing the reset password token
-   * @throws UserNotFoundException if the user with the provided email address is not found
-   * @throws ResetPasswordCodeInvalidException if the reset password code is invalid
-   * @throws ResetPasswordCodeExpiredException if the reset password code has expired
-   */
-  @Override
-  public InitiatePasswordChangeResponse verifyResetPasswordCode(final ResetPasswordDto resetPasswordDto) {
-    // Retrieve the email address from the reset password DTO
-    final String emailAddress = resetPasswordDto.getEmailAddress();
-    // Fetch the member by email address or throw an exception if not found
-    final Member member = getMemberDetails(emailAddress);
-
-    // Validate the provided verification code against the stored profile token and reset code
-    validateProfileTokenAndResetPasswordCode(emailAddress, resetPasswordDto.getVerificationCode());
-    // Initialize the user's authentication and security context
-    final FleenUser user = initializeAuthenticationAndContext(member);
-    // Generate a new reset password token for the user
-    final String resetPasswordToken = tokenService.createResetPasswordToken(user);
-
-    // Clear any temporarily saved OTP for the user
-    clearResetPasswordOtpSavedTemporarily(user.getUsername());
-    // Save the new reset password token for the user
-    tokenService.saveResetPasswordToken(user.getUsername(), resetPasswordToken);
-    // Return a localized response with the reset password token
-    return localizedResponse.of(InitiatePasswordChangeResponse.of(resetPasswordToken));
-  }
-
-  /**
    * Changes the password for the authenticated user.
    *
    * @param changePasswordDto the DTO containing the new password information
@@ -756,7 +689,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Check if user has associated reset password access token
     verifyUserHasResetPasswordToken(emailAddress);
     // Retrieve member from repository or throw exception if not found
-    final Member member = getMemberDetails(emailAddress);
+    final Member member = verificationService.getMemberDetails(emailAddress);
 
     // Find any existing password reset token and reset or clear details
     findPasswordTokenAndResetOrClearDetails(emailAddress);
@@ -839,39 +772,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     member.setCountry(country.getTitle());
   }
 
-  /**
-   * Initializes authentication and sets the authentication context for the provided member.
-   * Throws {@link FailedOperationException} if the member is null.
-   *
-   * @param member The member for whom authentication is to be initialized.
-   * @return The {@link FleenUser} associated with the member after authentication and context setup.
-   * @throws FailedOperationException If the member is null.
-   */
-  protected FleenUser initializeAuthenticationAndContext(final Member member) {
-    // Throw an exception if the provided member is null
-    checkIsNull(member, FailedOperationException::new);
-
-    // Create FleenUser from Member
-    final FleenUser user = FleenUser.fromMember(member);
-    // Create Authentication object
-    final Authentication authentication = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
-    // Set Authentication object in context (example: Spring Security context)
-    setContext(authentication);
-    // Return the authenticated user
-    return user;
-  }
-
-  /**
-   * Sets the authentication context in the security context holder if the authentication
-   * object is not null and is authenticated.
-   *
-   * @param authentication The authentication object to set in the security context holder.
-   */
-  protected void setContext(final Authentication authentication) {
-    if (nonNull(authentication) && authentication.isAuthenticated()) {
-      SecurityContextHolder.getContext().setAuthentication(authentication);
-    }
-  }
 
   /**
    * Saves the access token and refresh token for the given username using the JWT service.
@@ -1001,37 +901,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Compare the email addresses and throw an exception if they don't match
     if (!authenticatedUserEmailAddress.equals(requestEmailAddress)) {
       throw new FailedOperationException();
-    }
-  }
-
-  /**
-   * Clears the temporarily saved MFA (Multi-Factor Authentication) verification code
-   * from the cache associated with the given username.
-   *
-   * @param username the username associated with the MFA verification code to clear
-   */
-  private void clearMfaVerificationCodeSavedTemporarily(final String username) {
-    // Retrieve the cache key for the sign-up verification code
-    final String key = getMfaAuthenticationCacheKey(username);
-    // Delete the verification code from the cache
-    cacheService.delete(key);
-  }
-
-  /**
-   * Validates the Multi-Factor Authentication (MFA) verification code or OTP (One-Time Password) code
-   * based on the provided MFA type and username.
-   *
-   * @param otpCode  the MFA or OTP code to validate
-   * @param mfaType  the type of MFA (phone/email or authenticator app)
-   * @param username the username associated with the MFA or OTP code
-   */
-  private void validateMfaVerificationOrOtpCode(final String otpCode, final MfaType mfaType, final String username, final Long userId) {
-    if (mfaService.isPhoneOrEmailMfaType(mfaType)) {
-      // Validate email/phone MFA verification code
-      mfaService.validateEmailOrPhoneMfaVerificationCode(username, otpCode);
-    } else if (mfaService.isAuthenticatorMfaType(mfaType)) {
-      // Validate authenticator app MFA verification code
-      mfaService.validateAuthenticatorMfaVerificationCode(otpCode, userId);
     }
   }
 
@@ -1235,7 +1104,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Create an authentication token using the user's information and authorities
     final Authentication authenticationToken = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
     // Set the authentication context with the authentication token
-    setContext(authenticationToken);
+    verificationService.setContext(authenticationToken);
   }
 
   /**
@@ -1322,7 +1191,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Create an authentication token with the user's authorities
     final Authentication authenticationToken = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
     // Set the authentication token in the security context
-    setContext(authenticationToken);
+    verificationService.setContext(authenticationToken);
   }
 
   /**
@@ -1428,7 +1297,7 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     // Generate a refresh token for the user
     final String refreshToken = tokenService.createRefreshToken(user);
     // Set the authentication context with the provided authentication object
-    setContext(authentication);
+    verificationService.setContext(authentication);
     // Save the generated tokens for the user
     saveAuthenticationTokensToRepositoryOrCache(user.getUsername(), accessToken, refreshToken);
     // Update the sign-in response with the generated tokens and authentication status
@@ -1470,14 +1339,6 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     profileTokenRepository.save(profileToken);
   }
 
-  /**
-   * Clears the saved reset password OTP (One-Time Password) from the cache.
-   *
-   * @param username the username for which the OTP was saved
-   */
-  protected void clearResetPasswordOtpSavedTemporarily(final String username) {
-    cacheService.delete(getResetPasswordCacheKey(username));
-  }
 
   /**
    * Validates the profile token and reset password code for a given email address.
@@ -1489,57 +1350,13 @@ public class AuthenticationServiceImpl implements AuthenticationService,
    */
   private void validateProfileTokenAndResetPasswordCode(final String emailAddress, final String verificationTokenOrCode) {
     // Find the profile token associated with the email address
-    final ProfileToken profileToken = findProfileToken(emailAddress);
+    final ProfileToken profileToken = verificationService.findProfileToken(emailAddress);
     // Validate the provided verification token or code against the profile token
-    validateProfileTokenAndVerificationCode(verificationTokenOrCode, profileToken);
+    verificationService.validateProfileTokenAndVerificationCode(verificationTokenOrCode, profileToken);
     // Verify that the reset password token associated with the profile token has not expired
-    verifyResetPasswordTokenHasNotExpired(profileToken);
+    verificationService.verifyResetPasswordTokenHasNotExpired(profileToken);
   }
 
-  /**
-   * Finds the profile token associated with the given email address.
-   *
-   * @param emailAddress the email address associated with the profile token
-   * @return the profile token found
-   * @throws ResetPasswordCodeInvalidException if no profile token exists for the email address
-   */
-  protected ProfileToken findProfileToken(final String emailAddress) {
-    // Get the profile token and If no profile token exists, throw an exception indicating invalid reset password code.
-    return profileTokenRepository.findByEmailAddress(emailAddress)
-      .orElseThrow(ResetPasswordCodeInvalidException.of());
-  }
-
-  /**
-   * Validates the profile token and verification code.
-   *
-   * @param verificationTokenOrCode the verification token or code provided by the user
-   * @param profileToken the profile token containing the reset password token to validate against
-   * @throws ResetPasswordCodeInvalidException if the reset password token in the profile token is null
-   *                                          or does not match the provided verification token or code
-   */
-  protected void validateProfileTokenAndVerificationCode(final String verificationTokenOrCode, final ProfileToken profileToken) {
-    // Check if the provided token or code matches the reset password token in the profile token.
-    if (profileToken.isResetPasswordTokenInValid(verificationTokenOrCode)) {
-      throw new ResetPasswordCodeInvalidException();
-    }
-  }
-
-  /**
-   * Verifies that the reset password token has not expired.
-   *
-   * <p>This method checks if the provided {@code profileToken} exists and if its
-   * reset password token has expired. If the token has expired, it throws a
-   * {@link ResetPasswordCodeExpiredException}.</p>
-   *
-   * @param profileToken the profile token to be verified for expiry
-   * @throws ResetPasswordCodeExpiredException if the reset password token has expired
-   */
-  protected void verifyResetPasswordTokenHasNotExpired(final ProfileToken profileToken) {
-    // Check if the reset password token has expired.
-    if (nonNull(profileToken) && profileToken.isResetPasswordTokenExpired()) {
-      throw new ResetPasswordCodeExpiredException();
-    }
-  }
 
   /**
    * Finds the profile token by email address and resets the reset password token and its expiry date if found.
@@ -1631,18 +1448,5 @@ public class AuthenticationServiceImpl implements AuthenticationService,
     }
   }
 
-  /**
-   * Retrieves member details based on the provided email address or username.
-   *
-   * <p>This method looks up the member by their email address or username. If no matching member is found,
-   * a {@link UserNotFoundException} is thrown with the provided email address or username.</p>
-   *
-   * @param emailAddressOrUsername the email address or username of the member
-   * @return the {@link Member} object representing the member's details
-   * @throws UserNotFoundException if no member is found with the provided email address or username
-   */
-  protected Member getMemberDetails(final String emailAddressOrUsername) {
-    return memberRepository.findByEmailAddress(emailAddressOrUsername)
-      .orElseThrow(UserNotFoundException.of(emailAddressOrUsername));
-  }
+
 }
